Apri

+++ backend/app/Http/Controllers/CampaignController.php
@@ -12,10 +12,15 @@ use App\Models\User;
 use Carbon\Carbon;
 use DateTime;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
 
 class CampaignController extends Controller
 {
+
+    private $rek_destination = '082266605123';
+    private $contact_admin = '6282266605123';
+
     /**
      * disini nanti bakal ada pengaturan pengaturan
      * apakah perlu di filter yg bukan kedaluwarsa 
@@ -162,6 +167,7 @@ class CampaignController extends Controller
         }
     }
 
+    
     public function updateCampaign(Request $request, $id_campaign) {
 
         $this->validate($request, [
@@ -181,11 +187,13 @@ class CampaignController extends Controller
                 $image_name = Str::random(8) . date("Ymd");
                 
                 // cek apakah ada image yg perlu di hapus
-                $pos = strrpos($campaign->media_url, '/');
-                $curr_image = ($pos === false) ? $campaign->media_url : substr($campaign->media_url, $pos + 1);
-                $image_path = storage_path('uploads/image_campaign') . '/' . $curr_image;
-                if (file_exists($image_path)) {
-                    unlink($image_path);
+                if ($campaign->media_url) {
+                    $pos = strrpos($campaign->media_url, '/');
+                    $curr_image = ($pos === false) ? $campaign->media_url : substr($campaign->media_url, $pos + 1);
+                    $image_path = storage_path('uploads/image_campaign') . '/' . $curr_image;
+                    if (file_exists($image_path)) {
+                        unlink($image_path);
+                    }
                 }
                 
                 // simpan image
@@ -300,7 +308,7 @@ class CampaignController extends Controller
     private function generateNewTransaction($campaign, $user, $bank = 'OVO')
     {
         $price_after_fee = getCalculatedPrice($campaign->slot_price);
-        $random_code = mt_rand(10,999); // gunakan round aja kalau sistem ga support
+        $random_code = mt_rand(10,99); // gunakan round aja kalau sistem ga support
         $final_price = $price_after_fee + $random_code;
 
         $date = new DateTime();
@@ -310,6 +318,7 @@ class CampaignController extends Controller
                 'campaign_id' => $campaign->id, 
                 'user_id' => $user->id,
                 'bank' => $bank,
+                'no_rek_destination' => $this->rek_destination,
                 'no_transaction' => $date->getTimestamp(),
                 'type' => 1,
                 'timeout' => Carbon::now()->addHours(2),
@@ -318,7 +327,7 @@ class CampaignController extends Controller
                 'total_nominal' => $final_price,
             ]);
             // bank ini di kita nya
-            $url = "wa.me/628976634788?text=" . $user->id . '/' . $transaction->id . '/no_transaksi=' . $transaction->no_transaction;
+            $url = "wa.me/" . $this->contact_admin . "?text=" . $user->id . '/' . $transaction->id . '/no_transaksi=' . $transaction->no_transaction;
             $data = [
                 'name' => $user->name,
                 'desc' => $campaign->title,
@@ -410,5 +419,42 @@ class CampaignController extends Controller
             return response()->json(['message' => $e], 409);
         }
     }
+
+
+
+    public function cronCheckCampaign() {
+        try {
+            $campaigns = Campaign::with('campaign_members')
+                                 ->where('expired_date', '<', Carbon::now())->where('status', '0')
+                                 ->withCount(['campaign_members as slot_members' => function ($query) {
+                                    return $query->where('is_host', 0);
+                                }]);
+
+            if ($campaigns->doesntExist() || $campaigns->count() === 0) return response()->json(['message' => 'No Campaigns'], 200);
+            
+            // Sisanya tinggal dimasukin ke job, jadi langsung return, kalau ada yg gagal, ntar tinggal di masukin ke log
+            foreach ($campaigns->get() as $campaign) {
+                if (tap($campaign)->get()->slot_members === tap($campaign)->get()->slot_capacity) continue;
+                
+                $updated_campaign = $this->updateStatus($campaign, 2)->get();
+                $user = User::where('id', $updated_campaign->host_name->id)->first();
+                $type = "campaigns";
+                $emailJob = (new MailJob($user, $updated_campaign, $type));
+                dispatch($emailJob);
+            }
+
+            return response()->json(['message' => 'UPDATED'], 201);
+        } catch (\Exception $e) {
+            Log::error('Cron Update Transactions: ' . $e);
+            return response()->json(['message' => $e], 409);
+        }
+    }
+
+
+    private function updateStatus($campaign, $status)
+    {
+        return tap($campaign)->update(['status' => $status]);
+    }
+
 
 }
